Fix custom style sheet not loading in embed
